WIP Refactor enhancing Persona
